% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphid.R
\docType{data}
\name{aphid}
\alias{aphid}
\title{Aphid femur lengths}
\format{
A data frame with 384 rows and 3 variables:
\tabular{rl}{
 femur        \tab Femur length in 0.1 mm \cr
}
}
\usage{
data(aphid)
}
\description{
Twenty-five femur lengths of stem mothers of the aphid _Pemphigus
populitransversus_ in 0.1 mm.
}
\references{
Data from R. R. Sokal.
}
\keyword{datasets}
