% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\docType{data}
\name{cow}
\alias{cow}
\title{Ayrshire cow butterfat content}
\format{
A data frame with 120 rows and 1 variable:
\tabular{rl}{
 fat        \tab Interorbital length in mm \cr
}
}
\usage{
data(pigeon)
}
\description{
Butter fat content in percentages for 120 three-year-old Ayran cows selected
at random from a Canadian cattle breeder's stock record book
}
\references{
Sokal & Rohlf Biometry 3rd, p 59.
}
\keyword{datasets}
