% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housefly.R
\docType{data}
\name{housefly}
\alias{housefly}
\title{Wing lengths of houseflies}
\format{
A data frame with 100 rows and 1 variable:
\tabular{rl}{
 wing_length        \tab Wing length in 0.1 mm \cr
}
}
\usage{
data(houseflies)
}
\description{
Lengths of 100 wings of houseflies in 0.1 mm
}
\references{
Sokal, R. R., and P. E. Hunger. 1955. A morphometric analysis of
DDT-resistant and nonresistant housefly strains. Ann. Entomol. Soc. Amer.
48:499-507
}
\keyword{datasets}
