% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigeon.R
\docType{data}
\name{pigeon}
\alias{pigeon}
\title{Interorbital width of domestic pigeons}
\format{
A data frame with 40 rows and 1 variable:
\tabular{rl}{
 width        \tab Interorbital length in mm \cr
}
}
\usage{
data(pigeon)
}
\description{
40 measurements of interorbital width of a sample of domestic pigeons.
}
\references{
Olson, E. C., and R. L. Miller. 1958. Morphological Integration.
}
\keyword{datasets}
